+++ lawmaking/src/main/java/com/everyones/lawmaking/global/service/TokenService.java
@@ -7,6 +7,7 @@ import com.everyones.lawmaking.global.error.AuthException;
 import com.everyones.lawmaking.global.error.UserException;
 import com.everyones.lawmaking.global.jwt.AuthToken;
 import com.everyones.lawmaking.global.jwt.AuthTokenProvider;
+import com.everyones.lawmaking.global.util.AuthenticationUtil;
 import com.everyones.lawmaking.global.util.CookieUtil;
 import com.everyones.lawmaking.repository.TokenRepository;
 import com.everyones.lawmaking.repository.UserRepository;
@@ -37,15 +38,13 @@ public class TokenService {
 
     private final AuthTokenProvider tokenProvider;
 
-    private final HttpServletRequest httpServletRequest;
-
     private final AppProperties appProperties;
 
 
     private static final String PROVIDER = "provider";
 
 
-    public Map<String, String> issueToken(Authentication authentication ) {
+    public Map<String, String> issueToken(Authentication authentication) {
 
 
         PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
@@ -93,7 +92,7 @@ public class TokenService {
         return authTokenMap;
     }
 
-    public Map<String, String> reissueToken(String refreshTokenFromCookie ) throws AuthException{
+    public Map<String, String> reissueToken(String refreshTokenFromCookie) throws AuthException {
 
         // 쿠키에서 가져온 refresh token으로 token 행 가져오기
         var savedToken = tokenRepository.findTokenByRefreshToken(refreshTokenFromCookie)
@@ -151,42 +150,23 @@ public class TokenService {
         return authTokenMap;
     }
 
-    public void invalidateToken() {
-
-        //리프레시 토큰으로 사용자 id, provider 가져오기
-        var refreshTokenFromCookie = CookieUtil.getCookie(httpServletRequest, REFRESH_TOKEN);
-        if (refreshTokenFromCookie.isPresent()) {
-            AuthToken refreshToken = tokenProvider.convertAuthToken(refreshTokenFromCookie.get().getValue());
-            //사용자 조회해서 refresh 토큰 열 null 처리
-            tokenRepository.findTokenByRefreshToken(refreshToken.getToken())
-                    .ifPresent(tokenRepository::delete);
-
-        }
+    public void invalidateToken(Long userId) {
+        tokenRepository.deleteByUserId(userId);
     }
 
 
-    private static boolean hasAuthority(Collection<? extends GrantedAuthority> authorities, String authority) {
-        if (authorities == null) {
-            return false;
-        }
 
-        for (GrantedAuthority grantedAuthority : authorities) {
-            if (authority.equals(grantedAuthority.getAuthority())) {
-                return true;
-            }
-        }
-        return false;
-    }
+    public void logout(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse) {
 
-    public void logout(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse){
-
+        var userId = AuthenticationUtil.getUserId()
+                .orElseThrow(UserException.UserNotFoundException::new);
         HttpSession session = httpServletRequest.getSession(false); // 기존 세션 가져오기
         if (session != null) {
             session.invalidate();
         }
         var cookieDomain = appProperties.getAuth().getCookieDomain();
-        invalidateToken();
-        CookieUtil.deleteCookieForClient(httpServletRequest,httpServletResponse,ACCESS_TOKEN,cookieDomain);
+        invalidateToken(userId);
+        CookieUtil.deleteCookieForClient(httpServletRequest, httpServletResponse, ACCESS_TOKEN, cookieDomain);
         CookieUtil.deleteCookie(httpServletRequest, httpServletResponse, REFRESH_TOKEN);
         CookieUtil.deleteCookie(httpServletRequest, httpServletResponse, JSESSIONID);
