+++ lawmaking/src/main/java/com/everyones/lawmaking/global/util/AuthenticationUtil.java
@@ -1,8 +1,10 @@
 package com.everyones.lawmaking.global.util;
 
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 
+import java.util.Collection;
 import java.util.Optional;
 
 public class AuthenticationUtil {
@@ -13,5 +15,18 @@ public class AuthenticationUtil {
             return Optional.empty();
         }
             return Optional.of(Long.parseLong(userDetails.getUsername()));
+    }
+
+    private static boolean hasAuthority(Collection<? extends GrantedAuthority> authorities, String authority) {
+        if (authorities == null) {
+            return false;
+        }
+
+        for (GrantedAuthority grantedAuthority : authorities) {
+            if (authority.equals(grantedAuthority.getAuthority())) {
+                return true;
+            }
+        }
+        return false;
     }
 }
